finalisation synchronisation mdp
