+++ app/config/config.yml
@@ -69,7 +69,6 @@ swiftmailer:
     host:      "%mailer_host%"
     username:  "%mailer_user%"
     password:  "%mailer_password%"
-    port:      "%mailer_port%"
     spool:     { type: memory }
 
 fos_rest:
